Define labeled metrics externally
